+++ replication-voter/replication-voter.js
@@ -1,13 +1,102 @@
 module.exports = function (RED) {
-  function mode(myArray) {
-    return myArray.reduce(
-      (a, b, i, arr) =>
-        arr.filter((v) => JSON.stringify(v) === JSON.stringify(a)).length >=
-        arr.filter((v) => JSON.stringify(v) === JSON.stringify(b)).length
-          ? a
-          : b,
-      null
-    );
+  function mode(values, margin) {
+    let checked = [];
+    let modeValues = [];
+
+    let i;
+
+    for (i = 0; i < values.length; i++) {
+      if (checked.includes(values[i])) continue;
+
+      checked[checked.length] = values[i];
+
+      let currentValues = [];
+      let j;
+
+      if (margin != null) j = 0;
+      else {
+        currentValues[0] = values[i];
+        j = i + 1;
+      }
+
+      for (; j < values.length; j++) {
+        if (
+          margin != null &&
+          values[j] >= values[i] * (1 - margin) &&
+          values[j] <= values[i] * (1 + margin)
+        )
+          currentValues[currentValues.length] = values[j];
+        else if (values[j] === values[i])
+          currentValues[currentValues.length] = values[j];
+      }
+
+      if (currentValues.length > modeValues.length) {
+        modeValues = currentValues;
+      }
+    }
+
+    return modeValues;
+  }
+
+  function mean(values) {
+    let total = 0;
+    let i;
+
+    for (i = 0; i < values.length; i++) {
+      total = total + values[i];
+    }
+
+    return total / values.length;
+  }
+
+  function max(values) {
+    let maxValue = values[0];
+    let i;
+
+    for (i = 1; i < values.length; i++) {
+      if (values[i] > maxValue) maxValue = values[i];
+    }
+
+    return maxValue;
+  }
+
+  function min(values) {
+    let minValue = values[0];
+    let i;
+
+    for (i = 1; i < values.length; i++) {
+      if (values[i] < minValue) minValue = values[i];
+    }
+
+    return minValue;
+  }
+
+  function msgToSend(values, config) {
+    if (values.length == 0) return null;
+
+    if (
+      config.margin == null ||
+      config.margin == 0 ||
+      config.valueType === "string"
+    )
+      return values[0];
+
+    if (config.result === "mean") return mean(values);
+
+    if (config.result === "highest") return max(values);
+
+    if (config.result === "lowest") return min(values);
+
+    return null;
+  }
+
+  function createSameTypeArray(arr, type) {
+    let newArr = [];
+    let i;
+
+    for (i = 0; i < arr.length; i++) {
+      if (typeof arr[i] === type) newArr[newArr.length] = arr[i];
+    }
   }
 
   function setMajorityStatus(node, msg, done) {
@@ -59,32 +148,53 @@ module.exports = function (RED) {
 
     node.on("input", function (msg, send, done) {
       if (msg.payload.constructor === Array) {
-        let arrayMode = mode(msg.payload);
+        let sameTypeArray = createSameTypeArray(msg.payload, config.valueType);
 
-        let majorCount = msg.payload.filter(
-          (x) => JSON.stringify(x) == JSON.stringify(arrayMode)
-        ).length;
+        let modeValues;
 
-        msg.payload = arrayMode;
+        if (config.valueType === "number" && config.margin != null)
+          modeValues = mode(sameTypeArray, config.margin / 100);
+        else modeValues = mode(sameTypeArray, null);
 
-        sendOut(node, msg, done, majorCount >= parseInt(config.majority));
+        msg.payload = msgToSend(modeValues, config);
+
+        allValues = [];
+
+        sendOut(
+          node,
+          msg,
+          done,
+          modeValues.length >= parseInt(config.majority)
+        );
       } else if (
-        msg.payload.constructor === Number &&
+        (msg.payload.constructor === Number ||
+          msg.payload.constructor === String) &&
         typeof config.countInputs != "undefined"
       ) {
-        allValues.push(msg.payload);
+        if (typeof msg.payload === config.valueType)
+          allValues.push(msg.payload);
 
         if (allValues.length == config.countInputs) {
-          let arrayMode = mode(allValues);
+          let modeValues;
 
-          let majorCount = allValues.filter(
-            (x) => JSON.stringify(x) == JSON.stringify(arrayMode)
-          ).length;
+          if (
+            config.valueType === "number" &&
+            config.margin != null &&
+            config.margin != 0
+          )
+            modeValues = mode(allValues, config.margin / 100);
+          else modeValues = mode(allValues, null);
 
-          msg.payload = arrayMode;
+          msg.payload = msgToSend(modeValues, config);
+
           allValues = [];
 
-          sendOut(node, msg, done, majorCount >= parseInt(config.majority));
+          sendOut(
+            node,
+            msg,
+            done,
+            modeValues.length >= parseInt(config.majority)
+          );
         }
       } else {
         allValues = [];