+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All notable changes to this project will be documented in this file.
   consecutive messages, which allows achieving a more concise conclusion about the general flow. Sliding window
   size can be defined in the node's settings. #57
 - Action Delay: refactor discard strategy test #55
+- Enhancement of Replication Voter node by adding the possibility of using strings and using a margin when calculating the majority value.
+
 
 ### Week of 28/10/2020 - 03/11/2020